Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai==0.6.0
elevenlabs==1.12.0
Pillow==10.1.0
aiofiles==23.2.1